Mark resources as testing
