+++ rest/rest-resources/src/test/java/org/apache/brooklyn/rest/util/TestingHaMasterCheckResource.java
@@ -23,9 +23,8 @@ import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 
-@Path("/")
 @Produces(MediaType.APPLICATION_JSON)
-public class HaMasterCheckResource {
+public class TestingHaMasterCheckResource {
 
     @POST
     @Path("/server/shutdown")
