+++ rest/rest-server-jersey/src/test/java/org/apache/brooklyn/rest/util/TestingHaMasterCheckResource.java
@@ -18,27 +18,24 @@
  */
 package org.apache.brooklyn.rest.util;
 
-import javax.ws.rs.GET;
+import javax.ws.rs.POST;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 
-import org.apache.brooklyn.rest.filter.HaHotStateRequired;
-
-@Path("/ha/method")
+@Path("/")
 @Produces(MediaType.APPLICATION_JSON)
-public class HaHotStateCheckResource {
+public class TestingHaMasterCheckResource {
 
-    @GET
-    @Path("ok")
-    public String ok() {
-        return "OK";
+    @POST
+    @Path("/server/shutdown")
+    public void shutdown() {
+
     }
 
-    @GET
-    @Path("fail")
-    @HaHotStateRequired
-    public String fail() {
-        return "FAIL";
+    @POST
+    @Path("/ha/post")
+    public void post() {
     }
+
 }
